# api-gateway/docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bookstore API Gateway
  description: API Gateway for Bookstore Microservices
  version: 1.0.0
  contact:
    name: Bookstore Team

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Users
    description: User management
  - name: Books
    description: Book catalog
  - name: Categories
    description: Book categories
  - name: Authors
    description: Author information
  - name: Cart
    description: Shopping cart
  - name: Orders
    description: Order management
  - name: Blog
    description: Blog posts

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: api-gateway

  # Auth endpoints
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                full_name:
                  type: string
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation error

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        '401':
          description: Invalid credentials

  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info
        '401':
          description: Unauthorized

  # Books endpoints
  /api/books:
    get:
      tags: [Books]
      summary: Get all books
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of books

  /api/books/{id}:
    get:
      tags: [Books]
      summary: Get book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book details
        '404':
          description: Book not found

  /api/books/search:
    get:
      tags: [Books]
      summary: Full-text search books
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: highlight
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Search results
        '400':
          description: Empty query

  /api/books/bestsellers:
    get:
      tags: [Books]
      summary: Get bestseller books
      responses:
        '200':
          description: List of bestsellers

  # Categories endpoints
  /api/categories:
    get:
      tags: [Categories]
      summary: Get all categories
      responses:
        '200':
          description: List of categories

  # Authors endpoints
  /api/authors:
    get:
      tags: [Authors]
      summary: Get all authors
      responses:
        '200':
          description: List of authors

  /api/authors/{id}:
    get:
      tags: [Authors]
      summary: Get author by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author details
        '404':
          description: Author not found

  # Cart endpoints
  /api/cart:
    get:
      tags: [Cart]
      summary: Get user's cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart contents
        '401':
          description: Unauthorized

    post:
      tags: [Cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id, quantity]
              properties:
                book_id:
                  type: integer
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item added
        '401':
          description: Unauthorized

  # Orders endpoints
  /api/orders:
    get:
      tags: [Orders]
      summary: Get user's orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
        '401':
          description: Unauthorized

    post:
      tags: [Orders]
      summary: Create new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shipping_address]
              properties:
                shipping_address:
                  type: string
                payment_method:
                  type: string
                  enum: [cod, bank_transfer, momo]
      responses:
        '201':
          description: Order created
        '401':
          description: Unauthorized

  # Blog endpoints
  /api/blogs:
    get:
      tags: [Blog]
      summary: Get all blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of blog posts

  /api/blogs/{id}:
    get:
      tags: [Blog]
      summary: Get blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Blog post details
        '404':
          description: Post not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT